Reorganise interfaceToLammps
