+++ interfaceToLammps/USER-CFDDEM/pair_gran_hertzFix_history.cpp
@@ -61,7 +61,6 @@ void PairGranHertzFixHistory::compute(int eflag, int vflag)
   if (eflag || vflag) ev_setup(eflag,vflag);
   else evflag = vflag_fdotr = 0;
 
-  computeflag = 1;
   int shearupdate = 1;
   if (update->setupflag) shearupdate = 0;
