+++ interfaceToLammps/lib/cfddem/Makefile.lammps
@@ -1,0 +1,5 @@
+# Settings that the LAMMPS build will import when this package library is used
+#
+cfddem_SYSINC = 
+cfddem_SYSLIB = 
+nufeb_SYSPATH =
